// apps/qiita-trend-backend/src/types/itemsData.d.ts
/**
 *ItemsData
 */
export interface ItemsData {
  /**
   *body
   */
  body: string
  /**
   *id
   */
  id: string
  /**
   *likesCount
   */
  likesCount: number
  /**
   *private
   */
  private: boolean
  /**
   *reactionsCount
   */
  reactionsCount: number
  /**
   *stocksCount
   */
  stocksCount: number
  /**
   *tags
   */
  tags: string[]
  /**
   *title
   */
  title: string
  /**
   *url
   */
  url: string
  /**
   *pageViewsCount
   */
  pageViewsCount: number | null
}